Removed option related to use of different JSON serializers. Got all unit tests passing. Reverted public controllers that weren't meant to be recreated as management controllers.
