+++ examples/Umbraco.AuthorizedServices.TestSite/appsettings.Development.json
@@ -43,7 +43,6 @@
             "AuthorizationUrlRequiresRedirectUrl": true,
             "RequestTokenPath": "/assetbank-my-assets-test/oauth/token",
             "RequestTokenFormat": "FormUrlEncoded",
-            "JsonSerializer": "SystemTextJson",
             "ClientId": "",
             "ClientSecret": "",
             "Scopes": "",
@@ -71,7 +70,6 @@
             "AuthorizationUrlRequiresRedirectUrl": true,
             "RequestTokenPath": "/oauth/v1/token",
             "RequestTokenFormat": "FormUrlEncoded",
-            "JsonSerializer": "SystemTextJson",
             "ClientId": "",
             "ClientSecret": "",
             "Scopes": "oauth forms crm.objects.contacts.read crm.objects.contacts.write",
@@ -233,7 +231,6 @@
             "UseProofKeyForCodeExchange": true,
             "RequestTokenPath": "/oauth/token",
             "RequestTokenFormat": "FormUrlEncoded",
-            "JsonSerializer": "SystemTextJson",
             "ClientId": "",
             "ClientSecret": "",
             "Scopes": "",
@@ -249,7 +246,6 @@
             "UseProofKeyForCodeExchange": false,
             "RequestTokenPath": "/oauth2/get_token",
             "RequestTokenFormat": "FormUrlEncoded",
-            "JsonSerializer": "SystemTextJson",
             "ClientId": "",
             "ClientSecret": "",
             "Scopes": "",
@@ -265,7 +261,6 @@
             "UseProofKeyForCodeExchange": false,
             "RequestTokenPath": "/oauth2/token",
             "RequestTokenFormat": "FormUrlEncoded",
-            "JsonSerializer": "SystemTextJson",
             "ClientId": "",
             "ClientSecret": "",
             "Scopes": "",
@@ -281,7 +276,6 @@
             "UseProofKeyForCodeExchange": false,
             "RequestTokenPath": "/api/oauth2/token",
             "RequestTokenFormat": "FormUrlEncoded",
-            "JsonSerializer": "SystemTextJson",
             "ClientId": "",
             "ClientSecret": "",
             "Scopes": "email",
@@ -297,7 +291,6 @@
             "UseProofKeyForCodeExchange": false,
             "RequestTokenPath": "/api/oauth.v2.access",
             "RequestTokenFormat": "FormUrlEncoded",
-            "JsonSerializer": "SystemTextJson",
             "ClientId": "",
             "ClientSecret": "",
             "Scopes": "users:read.email users:read channels:read groups:read mpim:read im:read",
