+++ src/Umbraco.AuthorizedServices/Extensions/HttpContextExtensions.cs
@@ -1,12 +1,13 @@
 using Microsoft.AspNetCore.Http;
+using Umbraco.AuthorizedServices.Controllers;
 
 namespace Umbraco.AuthorizedServices.Extensions;
 
 internal static class HttpContextExtensions
 {
-    public static string GetOAuth2AuthorizedServiceRedirectUri(this HttpContext httpContext) => string.Empty;
-    //    => $"{httpContext.Request.Scheme}://{httpContext.Request.Host}/umbraco/api/{nameof(AuthorizedServiceResponseController).Replace("Controller", string.Empty)}/{nameof(AuthorizedServiceResponseController.HandleOAuth2IdentityResponse)}";
+    public static string GetOAuth2AuthorizedServiceRedirectUri(this HttpContext httpContext)
+        => $"{httpContext.Request.Scheme}://{httpContext.Request.Host}/umbraco/api/{nameof(AuthorizedServiceResponseController).Replace("Controller", string.Empty)}/{nameof(AuthorizedServiceResponseController.HandleOAuth2IdentityResponse)}";
 
-    public static string GetOAuth1AuthorizedServiceRedirectUri(this HttpContext httpContext) => string.Empty;
-    //    => $"{httpContext.Request.Scheme}://{httpContext.Request.Host}/umbraco/api/{nameof(AuthorizedServiceResponseController).Replace("Controller", string.Empty)}/{nameof(AuthorizedServiceResponseController.HandleOAuth1IdentityResponse)}";
+    public static string GetOAuth1AuthorizedServiceRedirectUri(this HttpContext httpContext)
+        => $"{httpContext.Request.Scheme}://{httpContext.Request.Host}/umbraco/api/{nameof(AuthorizedServiceResponseController).Replace("Controller", string.Empty)}/{nameof(AuthorizedServiceResponseController.HandleOAuth1IdentityResponse)}";
 }
