+++ src/Umbraco.AuthorizedServices/Services/Implement/AuthorizedRequestBuilder.cs
@@ -10,9 +10,9 @@ namespace Umbraco.AuthorizedServices.Services.Implement;
 
 internal sealed class AuthorizedRequestBuilder : IAuthorizedRequestBuilder
 {
-    private readonly JsonSerializerFactory _jsonSerializerFactory;
+    private readonly IJsonSerializer _jsonSerializer;
 
-    public AuthorizedRequestBuilder(JsonSerializerFactory jsonSerializerFactory) => _jsonSerializerFactory = jsonSerializerFactory;
+    public AuthorizedRequestBuilder(IJsonSerializer jsonSerializer) => _jsonSerializer = jsonSerializer;
 
     public HttpRequestMessage CreateRequestMessageWithOAuth2Token<TRequest>(
         ServiceDetail serviceDetail,
@@ -154,8 +154,7 @@ internal sealed class AuthorizedRequestBuilder : IAuthorizedRequestBuilder
             return null;
         }
 
-        IJsonSerializer jsonSerializer = _jsonSerializerFactory.GetSerializer(serviceDetail.Alias);
-        var serializedContent = jsonSerializer.Serialize(requestContent);
+        var serializedContent = _jsonSerializer.Serialize(requestContent);
         return new StringContent(serializedContent, Encoding.UTF8, "application/json");
     }
