--- src/Umbraco.AuthorizedServices/Services/Implement/SystemTextJsonSerializer.cs
@@ -1,20 +1,0 @@
-using System.Diagnostics.CodeAnalysis;
-using System.Text.Json;
-using Umbraco.Cms.Core.Serialization;
-
-namespace Umbraco.AuthorizedServices.Services.Implement;
-
-internal class SystemTextJsonSerializer : IJsonSerializer
-{
-    private readonly JsonSerializerOptions _jsonSerializerOptions;
-
-    public SystemTextJsonSerializer() => _jsonSerializerOptions = new JsonSerializerOptions { PropertyNamingPolicy = JsonNamingPolicy.CamelCase };
-
-    public string Serialize(object? input) => JsonSerializer.Serialize(input, _jsonSerializerOptions);
-
-    public T? Deserialize<T>(string input) => JsonSerializer.Deserialize<T>(input, _jsonSerializerOptions);
-
-    public T? DeserializeSubset<T>(string input, string key) => throw new NotSupportedException();
-
-    public bool TryDeserialize<T>(object input, [NotNullWhen(true)] out T? value) where T : class => throw new NotImplementedException();
-}
